fix intel
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@6979 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/TetrahedralMesh.cpp
@@ -830,7 +830,7 @@ unsigned TetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::GetContainingElementIndex(Chas
     // If it's in none of the elements, then throw
     std::stringstream ss; 
     ss << "Point [";
-    for(unsigned j=0; j<SPACE_DIM-1; j++)
+    for(unsigned j=0; (int)j<(int)SPACE_DIM-1; j++)
     {
         ss << testPoint[j] << ",";
     }
@@ -873,8 +873,8 @@ unsigned TetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::GetContainingElementIndexWithI
 
     // If it's in none of the elements, then throw
     std::stringstream ss; 
-    ss << "Point [";
-    for(unsigned j=0; j<SPACE_DIM-1; j++)
+    ss << "Point [";    
+    for(unsigned j=0; (int)j<(int)SPACE_DIM-1; j++)
     {
         ss << testPoint[j] << ",";
     }
